Moved SequenceSegment for content expansion
